feat: generate zsh shell completion script during build process using clap_complete
This generates the completions for the binary invocation during the compilation process.
This ensures that as long as the binary can build that zsh will be able to use the
autocomplete for the different commands and sub commands that exist.
