+++ Cargo.toml
@@ -1,21 +1,6 @@
-[package]
-name = "lifestuff"
-version = "0.2.2"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.71"
-clap = { version = "~4", features = ["derive"] }
-colored = "2.1.0"
-reqwest = { version = "0.12.2", features = ["blocking", "json"] }
-serde = { version = "1.0.162", features = ["derive"] }
-serde_json = "1.0.96"
-strum = { version = "0.26.2", features = ["derive"] }
-time = { version = "~0.3", features = ["macros", "formatting", "parsing"] }
-regex = "1.10.6"
-
-[dev-dependencies]
-dotenv = "0.15.0"
-mockito = "1.5.0"
+[workspace]
+resolver = "2"
+members = [
+    "lifestuff",
+    "lifestuff-types"
+]
