+++ lifestuff-types/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "lifestuff-types"
+version = "0.0.1"
+edition = "2021"
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { version = "~4", features = ["derive"] }
+strum = { version = "~0", features = ["derive"] }
+regex = "~1"
+anyhow = "~1"
