+++ lifestuff-types/src/conversions.rs
@@ -1,0 +1,18 @@
+use clap::{Args, Subcommand};
+
+pub mod area;
+pub mod distance;
+
+#[derive(Debug, Args)]
+pub struct Conversions {
+    #[command(subcommand)]
+    pub convert_type: ConversionOption,
+}
+
+#[derive(Subcommand, Debug)]
+pub enum ConversionOption {
+    /// Area Conversions
+    Area(area::AreaConversion),
+    /// Distance Conversions
+    Distance(distance::DistanceConversion),
+}
