+++ lifestuff-types/src/conversions/area.rs
@@ -1,0 +1,34 @@
+use clap::{Args, Subcommand, ValueEnum};
+use strum::Display;
+
+#[derive(Debug, Args)]
+pub struct AreaConversion {
+    #[arg(long, required = true)]
+    pub from: AreaUnits,
+    pub value: f64,
+    #[arg(long, required = true)]
+    pub to: Vec<AreaUnits>,
+}
+
+//noinspection SpellCheckingInspection
+#[derive(Subcommand, Debug, Display, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum)]
+pub enum AreaUnits {
+    #[clap(action=clap::ArgAction::SetTrue, alias = "a")]
+    /// Convert Using Acres
+    Acres,
+    #[clap(action=clap::ArgAction::SetTrue, name="inches", aliases = ["i","in"])]
+    /// Convert Using Square Inches
+    SqInches,
+    #[clap(action=clap::ArgAction::SetTrue, name="km", aliases =["sqkm"])]
+    /// Convert Using Square Kilometres
+    SqKilometres,
+    #[clap(action=clap::ArgAction::SetTrue, name="metres",aliases = ["sqm","m"])]
+    /// Convert Using Square Metres
+    SquareMetres,
+    #[clap(action=clap::ArgAction::SetTrue, name="miles", aliases = ["sqmi","mi"])]
+    /// Convert Using Square Miles
+    SquareMiles,
+    #[clap(action=clap::ArgAction::SetTrue, name="sqft", aliases = ["squarefeet","s"])]
+    /// Convert Using Square Feet
+    SquareFeet,
+}
