+++ lifestuff/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "lifestuff"
+version = "0.2.3"
+edition = "2021"
+build = "build.rs"
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.71"
+colored = "2.1.0"
+reqwest = { version = "~0", features = ["blocking", "json"] }
+serde = { version = "~1", features = ["derive"] }
+serde_json = "~1"
+strum = { version = "~0", features = ["derive"] }
+time = { version = "~0.3", features = ["macros", "formatting", "parsing"] }
+regex = "~1"
+lifestuff-types = { path = "../lifestuff-types" }
+clap_builder = "4.5.21"
+
+[dev-dependencies]
+dotenv = "0.15.0"
+mockito = "1.5.0"
+
+[build-dependencies]
+lifestuff-types = { path = "../lifestuff-types" }
+clap = { version = "~4", features = ["derive"] }
+clap_complete = "~4"
