+++ lifestuff/src/dateinfo/common.rs
