+++ lifestuff/src/ddg/generate.rs
