--- src/main.rs
@@ -1,62 +1,0 @@
-use anyhow::Result;
-use clap::{Parser, Subcommand};
-use strum::Display;
-
-mod conversions;
-mod currency;
-mod dateinfo;
-mod ddg;
-mod interest;
-mod mileage;
-
-#[derive(Parser, Debug)]
-#[clap(
-    author = "Me!",
-    about = "Lifestyle library!",
-    version = "0.2.0",
-    long_about = "Something Something Something Daaaarksiiiide"
-)]
-struct Cli {
-    #[arg(short, long, global=true, action=clap::ArgAction::SetTrue)]
-    verbose: bool,
-    #[command(subcommand)]
-    command: Commands,
-}
-
-#[derive(Subcommand, Display, Debug)]
-enum Commands {
-    /// Unit conversions
-    Convert(conversions::Conversions),
-    /// Date Operations
-    Dates(dateinfo::DateOperations),
-    /// Interest Calculations
-    Interest(interest::Interest),
-    /// Currency Conversion Operations
-    Currency(currency::Currency),
-    /// Mileage Calculations
-    Mileage(mileage::Mileage),
-    /// DuckDuckGo Address
-    DDG(ddg::DDGOperations),
-}
-
-fn main() -> Result<()> {
-    let cli: Cli = Cli::parse();
-    let verbose = cli.verbose;
-    if verbose {
-        println!("CLI Args: {:?}", cli);
-    }
-    let da_answer = match cli.command {
-        Commands::Convert(args) => conversions::perform_conversion(args),
-        Commands::Dates(args) => dateinfo::handle_date_operations(args, verbose),
-        Commands::Interest(args) => interest::handle_interest_calculations(args, verbose),
-        Commands::Currency(args) => currency::handle_currency_operations(args, verbose),
-        Commands::Mileage(args) => mileage::handle_mileage_operations(args, verbose),
-        Commands::DDG(args) => ddg::handle_ddg_operations(args, verbose),
-    };
-
-    if let Err(e) = da_answer {
-        println!("{e}");
-    }
-
-    Ok(())
-}
